Adding a new message server side.
This would be sent when all changes are in for a particular generation and the server has sorted through them.
This would trigger a new part of the client which would tick the autamata then for the new generation.
The automata would trigger the response, which the server would check.